provider: Update request parameters to suport MDS 0.3.x
Based on these changes [1], the request parameters to the trips
endpoint have changed. In addition the API now requires an Accept
header to determin wihch version the client is expecting to operate
in.

This change attempts at keeping the previous behaviour and only
modifying requests that are explicitly specified to be version 0.3

[1] https://github.com/openmobilityfoundation/mobility-data-specification/compare/0.2.x...0.3.x